+++ src/interfaces/iTrilaterationCoordinate.ts
@@ -1,0 +1,17 @@
+/**
+ *
+ */
+export interface iTrilaterationCoordinate {
+  /**
+   *
+   */
+  x: number;
+  /**
+   *
+   */
+  y: number;
+  /**
+   *
+   */
+  z: number;
+}