+++ src/interfaces/index.ts
@@ -54,3 +54,4 @@ export { iDeviceCluster } from './iDevicecluster';
 export { iDaytime } from './iDaytime';
 export { iBlockAutomaticHandler } from './iBlockAutomaticHandler';
 export { iTemperatureMeasurement } from './iTemperatureMeasurement';
+export { iTrilaterationCoordinate } from './iTrilaterationCoordinate';